After too long, getting my development up to date so I can demo my vgoptimize

+++ common-tools/cnuphys/magfield/src/main/java/cnuphys/magfield/ComponentZoomer.java
@@ -17,39 +17,37 @@ import javax.swing.JPopupMenu;
 import javax.swing.SwingUtilities;
 
 public class ComponentZoomer implements MouseListener, MouseMotionListener {
-	
+
 	private boolean _isDragging;
 	private Point _startPP;
 	private Point _endPP;
 	private Rectangle _oldXorRect;
 
-	//component being zoomed
+	// component being zoomed
 	private JComponent _component;
-	
-	//the zoomable, which is probable the same as the component
+
+	// the zoomable, which is probable the same as the component
 	private IComponentZoomable _zoomable;
-	
-	//the original world system
+
+	// the original world system
 	private Rectangle.Double _defaultWorld;
-	
+
 	public ComponentZoomer(IComponentZoomable zoomable) {
 		_zoomable = zoomable;
 		_component = _zoomable.getComponent();
-		
+
 		_component.addMouseListener(this);
 		_component.addMouseMotionListener(this);
-		
+
 		Rectangle.Double wr = _zoomable.getWorldSystem();
 		_defaultWorld = new Rectangle.Double();
 		_defaultWorld.setRect(wr.x, wr.y, wr.width, wr.height);
 	}
 
-
 	@Override
 	public void mouseClicked(MouseEvent e) {
 		System.out.println("Mouse Clicked");
 	}
-
 
 	@Override
 	public void mousePressed(MouseEvent e) {
@@ -60,44 +58,40 @@ public class ComponentZoomer implements MouseListener, MouseMotionListener {
 		System.out.println("Mouse Pressed button = " + e.getButton());
 	}
 
-
 	@Override
 	public void mouseReleased(MouseEvent e) {
 		if (_isDragging) {
 			_isDragging = false;
 			System.out.println("Dragging ended.");
 			_endPP = new Point(e.getX(), e.getY());
-			
-			int delX = (int)(Math.abs(_endPP.x - _startPP.x));
-			int delY = (int)(Math.abs(_endPP.y - _startPP.y));
-			
-			if ((delX > 4) &&  delY > 4) {
+
+			int delX = (Math.abs(_endPP.x - _startPP.x));
+			int delY = (Math.abs(_endPP.y - _startPP.y));
+
+			if ((delX > 4) && delY > 4) {
 				Rectangle r = rectangleFromPoints(_startPP, _endPP);
 				Rectangle.Double wr = new Rectangle.Double();
 				localToWorld(r, wr);
 //				System.out.println("Zoom rect: " + wr);
-				
+
 				_zoomable.setWorldSystem(wr);
 				_component.repaint();
 			}
-			
+
 			_startPP = null;
 			_endPP = null;
 			_oldXorRect = null;
-			
+
 		}
 	}
-
 
 	@Override
 	public void mouseEntered(MouseEvent e) {
 	}
 
-
 	@Override
 	public void mouseExited(MouseEvent e) {
 	}
-
 
 	@Override
 	public void mouseDragged(MouseEvent e) {
@@ -106,32 +100,31 @@ public class ComponentZoomer implements MouseListener, MouseMotionListener {
 			_isDragging = true;
 			return;
 		}
-		
+
 		Rectangle cr = rectangleFromPoints(_startPP, e.getPoint());
-		
+
 		Graphics g = _component.getGraphics();
-		
+
 		g.setColor(Color.black);
-		g.setXORMode(Color.white); 
-		
+		g.setXORMode(Color.white);
+
 		if (_oldXorRect != null) {
 			g.drawRect(_oldXorRect.x, _oldXorRect.y, _oldXorRect.width, _oldXorRect.height);
 		}
-		
+
 		g.drawRect(cr.x, cr.y, cr.width, cr.height);
 		_oldXorRect = cr;
 		g.dispose();
 	}
 
-
 	@Override
 	public void mouseMoved(MouseEvent e) {
 	}
-	
-	//popup trigger
+
+	// popup trigger
 	private void handleRightClick(MouseEvent e) {
 		System.out.println("Mouse Pressed Right Click");
-		
+
 		JPopupMenu popupMenu = new JPopupMenu();
 		popupMenu.setLightWeightPopupEnabled(false);
 
@@ -152,14 +145,12 @@ public class ComponentZoomer implements MouseListener, MouseMotionListener {
 		popupMenu.show(_component, e.getX(), e.getY());
 
 	}
-	
+
 	/**
 	 * Given two points, return the rectangle
 	 * 
-	 * @param p1
-	 *            one point
-	 * @param p2
-	 *            the other point
+	 * @param p1 one point
+	 * @param p2 the other point
 	 * @return the rectangle created from two points
 	 */
 
@@ -181,7 +172,7 @@ public class ComponentZoomer implements MouseListener, MouseMotionListener {
 		int y = Math.min(p1.y, p2.y);
 		return new Rectangle(x, y, w, h);
 	}
-	
+
 	public void localToWorld(Rectangle r, Rectangle.Double wr) {
 		Point p0 = new Point(r.x, r.y);
 		Point p1 = new Point(r.x + r.width, r.y + r.height);
